Homework L1/user_manager/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
coverage==7.7.1
debugpy==1.8.14
decorator==5.2.1
ecdsa==0.19.0
executing==2.2.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.0
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
passlib==1.7.4
platformdirs==4.3.8
pluggy==1.5.0
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.8.0
pytest==8.0.2
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-jose==3.4.0
python-multipart==0.0.9
pywin32==310
pyzmq==26.4.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.36.3
tornado==6.5
traitlets==5.14.3
typing_extensions==4.12.2
uvicorn==0.29.0
wcwidth==0.2.13
